requirements.txt
numpy
pandas
xarray
herbie-data
rich
matplotlib
requests
python-dotenv
